build: enable golangci-lint and fix reported errors (#1661)

+++ network/retrieval/retrieve_test.go
@@ -378,7 +378,7 @@ func getAllRefs(testData []byte) (storage.AddressCollection, error) {
 func getChunks(store chunk.Store) (chunks map[string]struct{}, err error) {
 	chunks = make(map[string]struct{})
 	for po := uint8(0); po <= chunk.MaxPO; po++ {
-		last, err := store.LastPullSubscriptionBinID(uint8(po))
+		last, err := store.LastPullSubscriptionBinID(po)
 		if err != nil {
 			return nil, err
 		}
